Classification Simple Dockarized Fast Api App With data validation/requirements.txt
fastapi
uvicorn
scikit-learn
joblib
nump